microservices/extractor/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
azure-core==1.39.0
azure-storage-blob==12.28.0
babel==2.18.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
courlan==1.3.2
cryptography==46.0.7
dateparser==1.4.0
fastapi==0.136.0
feedparser==6.0.12
h11==0.16.0
htmldate==1.9.4
idna==3.11
isodate==0.7.2
justext==3.0.2
lxml==6.1.0
lxml-html-clean==0.4.4
pycparser==3.0
pydantic==2.13.3
pydantic-core==2.46.3
python-dateutil==2.9.0.post0
pytz==2026.1.post1
regex==2026.4.4
requests==2.33.1
sgmllib3k==1.0.0
six==1.17.0
starlette==1.0.0
tld==0.13.2
trafilatura==2.0.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.44.0
